ward-quiz/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-docx==1.1.2
pydantic==2.10.4
